D

Documentation

This module defines, from P, a datatype constructor D for representing terms with presuppositions within a dynamic semantics and defines from it an instance of the Effect class of Orchard, Petricek, and Mycroft. The result is an implementation of the framework for presupposition of chapter 3 of Grove 2019.

newtype D e a

A datatype D for dynamic presupposition. D takes P and implicitly gives it ReaderT functionality so that it is sensitive to the discourse context, as well as (implicitly) StateT functionality so that it is sensitive to an incoming index.

Constructors

D

Fields

runD :: Int -> (InfoState -> InfoState) -> P e (a, Int)

(>>>=) :: SeqSplit e f (MonoidPlus e f) => D e a -> (a -> D f b) -> D (Plus D e f) b

A makeshift graded bind so as to provide a non-conflicting identifier for use outside the current module.

upD :: a -> D () a

The graded monad operator upD is just return.

downD :: (SeqSplit e (MonoidPlus f ()) (MonoidPlus e (MonoidPlus f ())), SeqSplit f () (MonoidPlus f ())) => D e (a -> b) -> D f a -> D (MonoidPlus e (MonoidPlus f ())) b

The graded monad operator downD is just ap.

joinD :: SeqSplit e f (MonoidPlus e f) => D e (D f b) -> D (MonoidPlus e f) b

The graded monad operator joinD is just join.

anaph :: NatWitness n -> a -> D (Insert (NatWitness n) a e) b -> D e b

We implement anaphora resolution via a function anaph, which takes a position to plug, a term with which to plug it, and the effectual meaning undergoing anaphora.

the :: Lift (Entity -> Bool) -> D (Lift Entity, ()) (Lift Entity)

A dynamic meaning for the.

(>@) :: (SeqSplit e (MonoidPlus f ()) (MonoidPlus e (MonoidPlus f ())), SeqSplit f () (MonoidPlus f ())) => D e InfoState -> D f InfoState -> D (MonoidPlus e (MonoidPlus f ())) InfoState

A function >@ for dynamic discourse update.

reset :: D e a -> D e a

We can reset meanings by plugging in the identity context id and plugging any reference to an incoming state with 0.

unWrap :: D () a -> Maybe a

A function to unWrap a meaning and see what value it might be holding (something in the Maybe monad).

checkForTruth :: D () InfoState -> Maybe Bool

A function checkForTruth to check whether or not a trivially effectual information state is Just True, Just False, or Nothing.