Dynamism

Description

We here define a module for the management of dynamic discourse entities, along with nondeterminism, as set out in Charlow 2014.

Documentation

type DynamicEntity m = StateT [Entity] m Entity

Define a type for dynamic entities.

startState :: [Entity]

Start from empty.

bind :: Monad m => DynamicEntity m -> DynamicEntity m

Charlow 2014: p. 75, Def. 3.16

makeDynEntity :: Monad m => m Entity -> DynamicEntity m

aDyn :: OnePlacePred -> DynamicEntity []

Charlow 2014: p. 47, Def. 2.33

pro :: Monad m => DynamicEntity m

Charlow 2014: p. 47, Def. 2.34

runFromStart :: StateT [Entity] [] a -> [(a, [Entity])]

faDyn :: StateT [Entity] [] (a -> b) -> StateT [Entity] [] a -> StateT [Entity] [] b

Charlow 2014: p. 45, Fact 2.11

baDyn :: StateT [Entity] [] a -> StateT [Entity] [] (a -> b) -> StateT [Entity] [] b

Charlow 2014: p. 45, Fact 2.11

checkDynForTruth :: StateT [Entity] [] Bool -> Bool

Let's have a way of checking dynamic, nondeterministic booleans for truth.