Framework.Grammar.Parser

Description

CKY-style parsing for CCG.

Documentation

forParse :: ([a] -> [b]) -> ParseAs a -> ParseAs b

interpret :: forall (p :: Project) m. Interpretation p m => Int -> Expr -> Chart m

CKY-style CCG expression interpreter, but with parses cached by substring identity instead of by span, for stronger memoization.

interpretations :: forall (p :: Project) m. (Interpretation p m, Eq m) => Expr -> Int -> ParseAs m

newtype ParseAs a

Constructors

ParseAs

Fields

getList :: [a]