Framework.Lambda.Terms

Description

We encode (untyped) λ-calculus, with constants, and including a definition of probabilistic programs.

Documentation

betaDeltaNormal :: DeltaRule -> Term -> Term

Beta normal forms, taking delta rules into account.

betaEtaNormal :: Term -> Term

Beta-eta normal forms.

betaNormal :: Term -> Term

Beta normal forms.

type Constant = Either String Double

Constants are indexed by either strings or real numbers.

dCon :: Double -> Term

Turn a Double into a constant.

type DeltaRule = Term -> Maybe Term

The type of Delta rules.

etaNormal :: Term -> Term

Eta normal forms.

freeVars :: Term -> [VarName]

fresh :: [Term] -> [VarName]

sCon :: String -> Term

Make arbitrarily typed constants.

subst :: VarName -> Term -> Term -> Term

Substitutions.

data Term

Untyped λ-terms. Types are assigned separately (i.e., "extrinsically").

Constructors

Var VarName

Con Constant

Lam VarName Term

App Term Term

TT

Pair Term Term

Pi1 Term

Pi2 Term

Return Term

Let VarName Term Term

(@@) :: Term -> Term -> Term

Abbreviations for application and pairing.

(&) :: Term -> Term -> Term

Abbreviations for application and pairing.