Framework.Lambda.Types

Description

Types and type inference are defined. Types feature variables, quantified at the top level for some limited polymorphism. Constants may also be polymorphic.

Documentation

arity :: Type -> Int

asTyped :: Sig -> (Term -> Term) -> Typed -> Typed

order :: Type -> Int

type Sig = Constant -> Maybe Type

Assign types to constants.

ty :: Sig -> Term -> Typed

Provide a term along with its principal type, given a signature.

tyEq :: TypeRed -> TypeRel

Determine if two types are semantically equivalent, taking type reductions into account.

tyEq' :: TypeRel

Semantic type equality, taking popable types into account.

data Type

Arrows, products, and probabilistic types, as well as (a) abstract type constructors, and (b) type variables for encoding polymorphism.

Constructors

Atom String

Type :→ Type infixr 5

Unit

Type :× Type infixl 6

P Type

TyCon String [Type]

TyVar String

data Typed

Typed terms (where not every term has a type).

Constructors

Typed

Fields

termOf :: Term

typeOf :: Maybe Type

unify :: Constr -> Maybe Constr

Arrive at a most general unifier, or fail if none exists.