InfoStates

Documentation

This module defines a type for information states, along with various operations on them, including update (>+) and dynamic implication (=>>).

newtype Set a

A type of sets so that we may construct sets of partial assignments (i.e., lists of entities).

Constructors

Setof

Fields

getList :: [a]

addToSet :: a -> Set a -> Set a

compact :: Eq a => Set a -> Set a

helpShowSet :: Show a => [a] -> String

isEmpty :: Eq a => Set a -> Bool

A predicate isEmpty to check whether or not a set is empty.

type InfoState = [Entity] -> Set [Entity]

A type InfoState, inhabited by information states. These take partial assignments (i.e., lists of entities) and return sets of such.

type LiftedEntity = [Entity] -> Entity

A type LiftedEntity, inhabited by functions from partial assignments (i.e., lists of entities) to entities.

true :: InfoState

An information state true, which returns the singleton set of any partial assignment (i.e., list of entities) it is fed.

isTrue :: InfoState -> Bool

A predicate on information states which checks whether or not they are true (by running them on the empty partial assignment).

(>+) :: InfoState -> InfoState -> InfoState

Information state update.

(=>>) :: InfoState -> InfoState -> InfoState

Dynamic implication.

type family Lift a where ...

A type family Lift for lifting arbitrary types into their information state variants.

Equations

Lift Bool = InfoState

Lift Entity = LiftedEntity

Lift (a -> b) = Lift a -> Lift b

class LiftPred a where

A class LiftPred with a single method lift for lifting predicates into their information state variants.

Minimal complete definition

lift

Methods

lift :: ([Entity] -> a) -> Lift a

dyn :: LiftPred a => a -> Lift a

From lift, we can define a function dyn which lifts predicates of any arity into their dynamic correspondents.