K

Documentation

This module defines, from D, a datatype constructor K for representing terms with presuppositions dynamically and in a continuized format so that one can analyze the semantic contribution of indefinites. We define from K an instance of the PMonad class of Orchard Petricek, and Mycroft. Although K constitutes a parameterized monad in the implementation, these gives rise to graded monads (in real life), given a monoid over the parameters. The result is thus a faithful implementation of the framework for presupposition of chapter 3 of Grove 2019.

newtype PContT m r f g a

A datatype constructor PContT for transforming a graded monad into its continuized variant as a parameterized monad.

Constructors

PContT

Fields

runPContT :: (a -> m g r) -> m f r

type K e f a = PContT D InfoState e f a

The type synonym K transforms D into its continuized variant.

(>>=>) :: Effect m => PContT m r f e a -> (a -> PContT m r e g b) -> PContT m r f g b

A makeshift parameterized bind to avoid naming conflicts outside the current module.

upK :: a -> K e e a

The parameterized monad operator upK is just return.

downK :: K f e (a -> b) -> K e g a -> K f g b

The parameterized monad operator downK is just ap.

joinK :: K f e (K e g b) -> K f g b

The parameterized monad operator joinK is just join.

monadicLift :: SeqSplit e f (MonoidPlus e f) => D e a -> K (MonoidPlus e f) f a

A monadicLift to transform objects in D to objects in K.

monadicLower :: K e () InfoState -> D e InfoState

A monadicLower to transform information states in K to information states in D.

newRegister :: InfoState

A function for generating a newRegister for use in the semantics of a.

a :: (SeqSplit () (MonoidPlus e ()) (MonoidPlus e ()), SeqSplit e () (MonoidPlus e ())) => Lift (Entity -> Bool) -> K (MonoidPlus e ()) e (Lift Entity)

A function a which provides the meaning of the indefinite article.

an :: (SeqSplit () (MonoidPlus e ()) (MonoidPlus e ()), SeqSplit e () (MonoidPlus e ())) => Lift (Entity -> Bool) -> K (MonoidPlus e ()) e (Lift Entity)

An allomorph of a for when it is phonotactically necessary.