Lambda.Types

Description

Types and type inference are defined. Types feature variables, quantified at the top level for some limited polymorphism. Constants may also be polymorphic.

Documentation

arity :: Type -> Int

asTyped :: Sig -> (Term -> Term) -> Typed -> Typed

data Atom

Atomic types for entities, truth values, and real numbers.

Constructors

E

T

R

order :: Type -> Int

type Sig = Constant -> Maybe Type

Assign types to constants.

ty :: Sig -> Term -> Typed

Provide a term along with its principal type, given a signature.

data Type

Arrows, products, and probabilistic types, as well as (a) abstract types representing the addition of a new Q, and (b) type variables for encoding polymorphism.

Constructors

At Atom

Type :→ Type infixr 5

Unit

Type :× Type infixl 6

P Type

Q Type Type Type

TyVar String

data Typed

Typed terms (where not every term has a type).

Constructors

Typed

Fields

termOf :: Term

typeOf :: Maybe Type

unify :: Constr -> Maybe Constr

Arrive at a most general unifier, or fail if none exists.