P

Documentation

This module defines a datatype constructor P for representing presupposition and defines from it an instance of the Effect class of Orchard, Petricek, and Mycroft. The result is an implementation of the framework for presupposition of chapter 2 of Grove 2019.

data Seq p where

Seq is datatype constructor for heterogeneous sequences. The type of its argument is the type of the effect. The resulting datatype comes with functions head and tail for extracting the head and tail of a sequence, respectively.

Constructors

End :: Seq ()

(:+) :: a -> Seq p -> Seq (a, p)

class HelpShowSeq a where

Minimal complete definition

helpShowSeq

Methods

helpShowSeq :: a -> String

offset :: String -> Int

takeFirstOf0 :: ([a] -> Bool) -> [a] -> [a] -> [a]

takeFirstOf :: ([a] -> Bool) -> [a] -> [a]

dropFirstOf0 :: ([a] -> Bool) -> [a] -> [a] -> [a]

dropFirstOf :: ([a] -> Bool) -> [a] -> [a]

class HelpReadSeq a where

Minimal complete definition

helpReadSeq

Methods

helpReadSeq :: String -> [(a, String)]

head :: Seq (a, p) -> a

A function for extracting the head of a sequence.

tail :: Seq (a, p) -> Seq p

A function for extracting the tail of a sequence.

type family MonoidPlus a b where ...

A type family corresponding to a monoid whose addition operation is MonoidPlus and whose unit is ().

Equations

MonoidPlus () a = a

MonoidPlus (a, p) b = (a, MonoidPlus p b)

(+:+) :: Seq a -> Seq b -> Seq (MonoidPlus a b)

Concatentation for sequences.

data Nat

A datatype for Naturals is used at the type level.

Constructors

Zero

Succ Nat

data NatWitness n where

Witnesses to the type level Naturals provide parameters for the function preAnaph.

Constructors

ZeroW :: NatWitness Zero

SuccW :: NatWitness n -> NatWitness (Succ n)

newtype P e a

A datatype for terms with presuppositions. Its first parameter is the effect, and its second is the type of the value.

Constructors

P

Fields

runP :: Seq e -> Maybe a

type family Insert i a e where ...

The type-level function to help type the function anaph.

Equations

Insert (NatWitness Zero) a p = (a, p)

Insert (NatWitness (Succ n)) a (b, p) = (b, Insert (NatWitness n) a p)

insert :: NatWitness n -> a -> Seq e -> Seq (Insert (NatWitness n) a e)

A function for inserting terms into sequences, used in the definition of anaph.

preAnaph :: NatWitness n -> a -> P (Insert (NatWitness n) a e) b -> P e b

A function to implement anaphora resolution for the graded monad P.

class SeqSplit s t st where

A class with a method seqSplit for splitting sequences in a way that depends on the desired type of the result. This trick echoes the method defined by Orchard, Petricek, and Mycroft in their definition of the graded Reader monad. The main difference is that, here, the effects are given by sequences, whereas their effects are given by sets.

Minimal complete definition

seqSplit

Methods

seqSplit :: Seq st -> (Seq s, Seq t)

(>>>>=) :: Maybe a -> (a -> Maybe b) -> Maybe b

Let's quickly redefine a >>>>= for Maybe, the original bind of which has been hidden.

upP :: a -> P () a

The graded monad operator upP is just return.

downP :: (SeqSplit e (MonoidPlus f ()) (MonoidPlus e (MonoidPlus f ())), SeqSplit f () (MonoidPlus f ())) => P e (a -> b) -> P f a -> P (MonoidPlus e (MonoidPlus f ())) b

The graded monad operator downP is just ap.

joinP :: SeqSplit f g (MonoidPlus f g) => P f (P g b) -> P (MonoidPlus f g) b

The graded monad operator joinP is just join.

(||-) :: Bool -> a -> Maybe a

The turnstile.

theSta :: (Entity -> Bool) -> P (Entity, ()) Entity

Let's define a static meaning for the.