Scope

Description

We here define a module for combining the management of dynamic discourse entities with scope via delimited continuations, as set out in Charlow 2014.

Documentation

type ScopalEntity m = ContT Bool m Entity

Define a type for scopal entities.

monadicLift :: Monad m => m a -> ContT r m a

Charlow 2014: p. 68, Def. 3.12

neg :: StateT [Entity] [] Bool -> StateT [Entity] [] Bool

Charlow 2014: p. 51, Def. 3.2

every :: (Entity -> Bool) -> ScopalEntity (StateT [Entity] [])

Charlow 2014: p. 54, Def. 3.4

no :: (Entity -> Bool) -> ScopalEntity (StateT [Entity] [])

Charlow 2014: p. 54, Def. 3.4

checkContForTruth :: ContT Bool (StateT [Entity] []) Bool -> Bool

Let's have a way of checking scopal booleans for truth.